move notification logic to alert view controller
