+++ DeFuge/AlarmsListViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import UserNotifications
 
 class AlarmsListViewController: UIViewController,UITableViewDataSource,UITableViewDelegate {
 
@@ -16,6 +17,10 @@ class AlarmsListViewController: UIViewController,UITableViewDataSource,UITableVi
         // Do any additional setup after loading the view, typically from a nib.
         tableView.dataSource = self
         tableView.delegate = self
+        
+        let alarm = Alarm()
+        alarm.time = Date().addingTimeInterval(20.0)
+        addAlarm(alarm: alarm)
     }
 
     override func didReceiveMemoryWarning() {
@@ -24,7 +29,7 @@ class AlarmsListViewController: UIViewController,UITableViewDataSource,UITableVi
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        var date = Date();
+        let date = Date();
         let cell = tableView.dequeueReusableCell(withIdentifier: "AlarmListCell") as! AlarmListCell
         
         let dateFormatter = DateFormatter()
@@ -35,7 +40,7 @@ class AlarmsListViewController: UIViewController,UITableViewDataSource,UITableVi
         let newFireDate = dateFormatter.string(from: fireDate)
         print(newFireDate) //New formatted Date string
         
-        setupNotification(date)
+//        setupNotification(date)
         cell.alarmLabel.text = newDate
         cell.alarmDesc.text = "Alarm"
         cell.alarmSwitch.isOn = true
@@ -52,6 +57,38 @@ class AlarmsListViewController: UIViewController,UITableViewDataSource,UITableVi
         AlarmNotification.soundName = "trial.mp3"
         AlarmNotification.timeZone = TimeZone.current
         UIApplication.shared.scheduleLocalNotification(AlarmNotification)
+    }
+    
+    
+    func addAlarm(alarm: Alarm){
+        // ask permission for notification
+        let center = UNUserNotificationCenter.current()
+        center.requestAuthorization(options: [.alert, .sound]) { (authorized, error) in
+            if authorized {
+                print("access granted, proceed")
+                var date = DateComponents()
+                let calendar = Calendar.current
+                
+                let hour = calendar.component(.hour, from: alarm.time)
+                let minutes = calendar.component(.minute, from: alarm.time)
+                let seconds = calendar.component(.second, from: alarm.time)
+                
+                date.hour = hour
+                date.minute = minutes
+                date.second = seconds
+                let trigger = UNCalendarNotificationTrigger(dateMatching: date, repeats: false)
+                
+                let content = UNMutableNotificationContent()
+                content.categoryIdentifier = "defugue"
+                content.title = "Alarm"
+                content.body = "wakey wakey"
+                content.userInfo = ["customNumber": 100, "time": "hours = \(hour):\(minutes):\(seconds)"]
+                content.sound = UNNotificationSound(named: "Elegant.mp3")
+                
+                let request = UNNotificationRequest(identifier: "exampleNotification", content: content, trigger: trigger)
+                center.add(request, withCompletionHandler: nil)
+            }
+        }
     }
