+++ DeFuge/Base.lproj/Main.storyboard
@@ -64,7 +64,9 @@
                         </subviews>
                         <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                     </view>
-                    <navigationItem key="navigationItem" title="Alarm" id="Saj-ua-BxL"/>
+                    <navigationItem key="navigationItem" title="Alarm" id="Saj-ua-BxL">
+                        <barButtonItem key="rightBarButtonItem" title="+" id="J1D-1d-2LM"/>
+                    </navigationItem>
                     <connections>
                         <outlet property="tableView" destination="lKg-i3-Ct3" id="6wv-hx-W46"/>
                     </connections>
